clean up packages move files
