+++ integration/src/test/scala/org/knora/webapi/core/LayersTest.scala
@@ -46,9 +46,9 @@ import org.knora.webapi.slice.ontology.repo.service.PredicateRepositoryLive
 import org.knora.webapi.slice.resourceinfo.ResourceInfoLayers
 import org.knora.webapi.slice.resourceinfo.api.service.RestResourceInfoService
 import org.knora.webapi.slice.resourceinfo.domain.IriConverter
-import org.knora.webapi.slice.search.search.api.SearchApiRoutes
-import org.knora.webapi.slice.search.search.api.SearchEndpoints
-import org.knora.webapi.slice.search.search.api.SearchEndpointsHandler
+import org.knora.webapi.slice.search.api.SearchApiRoutes
+import org.knora.webapi.slice.search.api.SearchEndpoints
+import org.knora.webapi.slice.search.api.SearchEndpointsHandler
 import org.knora.webapi.store.cache.CacheServiceRequestMessageHandler
 import org.knora.webapi.store.cache.CacheServiceRequestMessageHandlerLive
 import org.knora.webapi.store.cache.api.CacheService
